WL-1046 Recurring events in the signup tool are rendered as expanded by default.
This will probably be appreciated by novice users of the tool, less so by
intermediate to expert users as this quickly becomes unreadable for overlapping
recurring events. I tried to introduce a concession whereby only single
recurring events where rendered as expanded, but I couldn't find a place to put
the logic without incurring mutual recursion and / or incompatibilities with UI
widget state, so I just resorted to the initial request in the end!