WL-2347 create a site calendar if it doesn't exist yet
The signup tool currently depends on the calendar for a site being already
initialised by the calendar tool. When the site calendar is initialised by
adding the tool it checks to see if the calendar is already defined for
that site and if it is, it opts to use that one.

In fact when the calendar tool is removed, the site calendar is left there.
So there is little harm in initialising the calendar early, namely once the
signup event is requested.