WL-3094 Surpress a NullPointerException when emailing users
When you delete a sign up event an email will be sent out to notify users.
After lots of thinking about which users to send the email to, sometimes
it couldn't think of any. Once the `NullPointerException` had bubbled its
way back to the surface it was hard to tell whether the event had been
succesfully deleted, so it showed an error message, even though the deletion
was succesful.