WL-3110 Add server config for attendee emails
Currently an attendee doesn't receive an email if an event isn't set out
to send out notifications for all events. Arguably an attendee should get
an email when they get signed up, but it is a bit spammy to send out emails
everytime an event is modified.

As such we are introducing a new config variable for `sakai.properties`:

    signup.email.notification.attendee.signed.up

The default behaviour is, "yes do it please".