WL-3119 Ensure a event creator is defined
This ensures that an attendee gets assigned to the event, if this doesn't
happen the calendar will be invalid because it uses the `method=REQUEST`.

Infuriatingly the setCreator method on an event takes no parameters and
will assign the current user. This is not necessarily the creator of the
event.

Fortunately the event exposes its properties so we can just fiddle with
them directly. What could go wrong?