WL-3057 Change the group name prefix to S:
We have a prefix to distinguish groups that are defined by signup from
regular groups. We will be making these group names much longer so in the
good name of not using all caps, let's simplify things.

+ **Old style group name:** `SIGNUP_Election`
+ **New style group name:** `S:Election`

When synchronising these names we make the signup tool aware of the old
style. Otherwise it will think that a user has editted the group name
directly in Site Info.