WL-3119 Define MIME headers on Attachment
Use the new constructor in the kernel for Attachment that allows you
to specify the MIME headers, namely the `Content-Type` and
`Content-Disposition`.

This is important so that we can get the email client to recognise
that this is a calendar invite and we would like to display it as an
inline meeting request.

This is tied to a change in the kernel. Locally commit 34ae2f134e.