WL-3119 Set flag for multipart subtype
The Signup email façade needs to tell the `BasicEmailService` that the
message should have a different MIME multipart subtype because it is a
meeting invite.

A mechanism was introduce to the kernel in 27fee06f1a48965f

This involves adding the flag called `multipart-subtype` as one of the
'additional headers'. We sneak it in this way because the existing API
is quite restrictive and this will fail silently if it is not implemented
in the currently running version of kernel.