WL-3119 Meeting coordinators chair meetings
Meeting coordinators are added to the meeting event as chair participants.

Relies on the ExternalCalendaringService API being extended to define this
specific role. Normally participants are added as 'required participants'.

I was going to define something similar for the attendees but this would
be a privacy concern. Better to add them later in the EmailFacade when the
recipient is known.