WL-3170 Attendees won't cancel meetings
An ical file is sent to an organiser when a signup is cancelled by an
attendee to update the list of the attendees of the meeting. This should
not be a cancellation message because the meeting has not been cancelled.

The text of this email is quite confusing and should probably be fixed too.