WL-3170 Cancel? Pick the right event
When an attendee cancels their own signup the event is not picked up in
the code as it stands. Instead of looking for events that the user is
attending we should be using the internal list of removed events for the
class.

Of course the list will only have one element but how would the interface
know that now?