SIGNUP-243 Make ICS attachments work for web browser email clients
This refactors the behaviour for setting up different types of email to their own classes,
and introduces some testing. This makes it easier to implement aspects of the iCal and iTIP
specifications such as cancellation of events.

The new code uses changes in the kernel to specify extra information about the email attachments
to indicate to the mail client what type of event is being described.